requirements.txt
pytelegrambotapi
openai
requests
beautifulsoup4
python-dotenv
